+++ src/base/index.ts
@@ -1,1 +1,6 @@
 export { MongoDB } from './mongodb';
+export {
+  HttpServer,
+  HttpServerCreateOptions,
+  HttpServerInstance,
+} from './httpServer';